requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.15.4
frozenlist==1.4.1
idna==3.7
jieba3k==0.35.1
joblib==1.4.2
JPype1==1.5.0
konlpy==0.6.0
lxml==5.2.2
lxml_html_clean==0.2.0
multidict==6.0.5
newspaper3k==0.2.8
nltk==3.8.1
numpy==2.0.1
packaging==24.1
pillow==10.4.0
pip==24.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
requests-file==2.1.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
tinysegmenter==0.3
tldextract==5.1.2
tqdm==4.66.4
urllib3==2.2.2
yarl==1.9.4
